==> 2. Descriptive Analysis.Rmd <==
---
title: "Descriptive Analysis"
output: html_document

knit: (function(input_file, encoding) {
  out_dir <- 'docs';
  rmarkdown::render(input_file,
 encoding=encoding,
 output_file=file.path(dirname(input_file), out_dir, 'Descriptive_analysis.html'))})
---

#### Define Libraries
```{r echo=TRUE, message=FALSE, warning=FALSE, tidy=TRUE}
library("stringr")
library("dplyr")
library("ggplot2")
library("gridExtra")
library("knitr")
library("tidyverse")
library("magrittr")
library("foreign")
library("AER")
library("ggpubr")
library("scales")
library("devtools")
library("weights")
library("cowplot")
library("usmap")
library("RColorBrewer")
library("viridis")
```

#### Define Path
```{r tidy=TRUE}
dir.wrk <- getwd()
dir.data <- file.path(dir.wrk, "dataset/data")
dir.outdata <- file.path(dir.wrk, "dataset/outputdata")
dir.plot <- file.path(dir.wrk, "dataset/plots_descriptive")
```

#### Define Files and load data (created in data preparation)
```{r tidy=TRUE}
file.dat <- file.path(dir.outdata, "dat_ineq.csv")
dat_p <- read.csv(file.dat, header=TRUE, stringsAsFactors=FALSE)
```

```{r setup, include=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

#### Population-weighted per capita CO2 by party (metric tonnes)
```{r tidy=TRUE}

## weighted mean CO2 by Party and Year
dat_mean <- dat_p %>% group_by(Party, Year) %>% summarise(pw_mean=weighted.mean(pc_CO2, pop_mi))
dat_mean_spread <- dat_mean %>% spread(Party, pw_mean)

## weighted national average CO2 by year 
Nat_Ave_CO2 <- dat_p %>% group_by(Year) %>% summarise(nat_ave=weighted.mean(pc_CO2, pop_mi))

## join flies: weighted mean CO2 by party
party_co2 <- list(Nat_Ave_CO2, dat_mean_spread) %>% reduce(left_join, by = c("Year"))
party_co2_long <- party_co2 %>% gather(Party, mvalue, -Year)
party_co2_long <- party_co2_long %>%  mutate(Party = fct_relevel(Party, "Rep", "nat_ave", "Dem"))

## output data
file.output <- file.path(dir.outdata, "mean_co2_party.csv")
write.table(party_co2, file.output, sep=",", row.names=FALSE, col.names=TRUE, quote=FALSE)
```
</br>

#### Plot: Mean CO2 Red, Blue and weighted Average
```{r tidy=TRUE}
co2_p <- ggplot(party_co2_long, aes(x=Year))+
  geom_line(data = party_co2_long, aes(y=mvalue, color= factor(Party)), linewidth = 0.7)+
  scale_color_manual(labels=c("Red states", "National average", "Blue states"), values = c("red", "green", "blue"))+
  ylab("")+
  xlab("")+
  scale_x_continuous(breaks=c(1997, 2000, 2003, 2006, 2009, 2012, 2015, 2018, 2021)) +
  theme(axis.text.x = element_text(angle = 90, vjust=0.5))+
  theme(axis.text=element_text(size=7))+
  theme(axis.title.x=element_text(size = 8))+
  theme(legend.title = element_blank())+
  guides(linetype = guide_legend(reverse = TRUE))+
  scale_y_continuous(limits = c(8, 28), breaks = seq(8, 28, 2)) +
  theme(legend.position = c(0.32, 0.21))+
  theme(legend.key.width = unit(0.5, "cm"))+
  theme(legend.key.height = unit(0.4, "cm"))+
  theme(legend.key.size= unit(0.5, "cm")) +
  theme(legend.direction = "vertical") +
  theme(legend.background = element_rect(fill = "transparent")) +
  theme(legend.spacing.x = unit(0.3, "lines"))+
  theme(legend.spacing.y = unit(0.8, "lines"))+
  theme(legend.text=element_text(size=8))
```

</br>
#### Population-weighted per capita CO2 by fuel type and party (metric tonnes)
```{r fig.height = 3, fig.width = 5, fig.align = "center"}

d_fuel <- dat_p %>% group_by(Party, Year) %>%
  summarise(pw_coal=weighted.mean(pc_coal, pop_mi), 
            pw_petro=weighted.mean(pc_petro, pop_mi),
            pw_gas=weighted.mean(pc_gas, pop_mi))

dfuel_year <- dat_p %>% group_by(Year) %>%
  summarise(pw_coal=weighted.mean(pc_coal, pop_mi), 
            pw_petro=weighted.mean(pc_petro, pop_mi),
            pw_gas=weighted.mean(pc_gas, pop_mi))
  
## Population-weighted per capita CO2 by Fuel Type and Party 
d_fuel <- d_fuel %>% gather(fuel_type, mvalue, -c(Party, Year))
d_fuel <- d_fuel %>%  mutate(fuel_type = fct_relevel(fuel_type, "pw_coal", "pw_petro", "pw_gas"))
d_fuel <- d_fuel %>%  mutate(Party = fct_relevel(Party, "Rep", "Dem"))

## Facet Wrap
d_fuel$fuel_type <- factor(d_fuel$fuel_type, labels = c("Coal", "Petroleum", "Natural Gas"))

## plot
fuel_p <- ggplot(data = d_fuel, aes(x = Year, y = mvalue, color=factor(Party))) +
  geom_line(aes(y=mvalue), size =0.6) + 
  geom_point(aes(y=mvalue), size = 0.5) +
  scale_x_continuous(breaks=c(1997, 2000, 2004, 2008, 2012, 2016, 2019, 2021)) +
  scale_y_continuous(limits=c(0, 12), breaks=seq(0, 12, 2)) +
  scale_colour_manual(name = "States", labels = c("Blue States", "Red States"),
                      values=c("blue", 'red'), guide = guide_legend(reverse=T))+
  labs(x = "", y = "")+
  theme(axis.title.x = element_text(size=8))+
  theme(axis.text.x = element_text(angle = 90, vjust=0.5))+
  theme(axis.text= element_text(size=7))+
  facet_wrap(~fuel_type, nrow=1, ncol=5, scales="fixed", drop=TRUE, strip.position="top")+
  theme(legend.position ="none")+
  theme(legend.direction = "horizontal") +
  #theme(legend.position = "bottom", legend.margin=margin(t=-5)) +
  theme(legend.box = "vertical")+
  theme(legend.title = element_blank()) +
  theme(legend.key.width = unit(1, "cm"))+
  theme(legend.spacing.y = unit(0.1, "cm"))+
  theme(legend.text = element_text(size=8))
```
</br>

#### Merge plots (Overall and fuel type)
```{r tidy=TRUE}
co2_fuel <- plot_grid(co2_p, fuel_p, nrow=1, ncol=2, labels = c("a", "b"), label_size = 8, color = "black", face = "bold", hjust=-2, rel_widths= c(0.6, 1))
```
</br>

#### Population-weighted per capita CO2 by SECTOR and party (metric tonnes)
```{r fig.height = 3, fig.width = 5, fig.align = "center"}

d_sector <- dat_p %>% group_by(Party, Year) %>% 
  summarise(pw_comm=weighted.mean(pc_comm, pop_mi),
            pw_res=weighted.mean(pc_res, pop_mi),
            pw_ind=weighted.mean(pc_ind, pop_mi),
            pw_elec=weighted.mean(pc_elec, pop_mi),
            pw_trans=weighted.mean(pc_trans, pop_mi))

sector_s <- d_sector %>% gather(sector, mvalue, -c(Party, Year))
sector_s <- sector_s %>%  mutate(sector = fct_relevel(sector, "pw_elec", "pw_trans", "pw_ind",
                                                      "pw_res", "pw_comm"))
## Facet Wrap
sector_s$sector <- factor(sector_s$sector, labels = c("Electricity", "Transportation", "Industrial",
                                                      "Residential", "Commerical"))
## plot
sector_p <- ggplot(data = sector_s, aes(x = Year, y = mvalue, color=factor(Party))) +
  geom_line(aes(y=mvalue), size =0.6) + # CHANGED COLOR HERE
  geom_point(aes(y=mvalue), size = 0.5) + # CHANGED COLOR HERE
  scale_x_continuous(breaks=c(1997, 2000, 2004, 2008, 2012, 2016, 2019, 2021)) +
  theme(axis.text= element_text(size=7))+
  scale_y_continuous(breaks=seq(0, 12, 2)) +
  scale_colour_manual(name = "States", labels = c("Blue States", "Red States"),
                      values=c("blue", 'red'), guide = guide_legend(reverse=TRUE))+
  theme(axis.text.x = element_text(angle = 90, vjust=0.5))+
  theme(axis.text=element_text(size=7))+
  facet_wrap(~sector, nrow=1, ncol=5, scales="fixed", drop=TRUE, strip.position="top")+
  labs(x = "Year", y = "")+
  theme(axis.title=element_text(size=10))+
  theme(legend.direction = "horizontal") +
  theme(legend.position = "bottom", legend.margin=margin(t=-5)) +
  #theme(legend.position ="none")+
  theme(legend.box = "vertical")+
  theme(legend.title = element_blank()) +
  theme(legend.key.width = unit(1, "cm"))+
  theme(legend.spacing.y = unit(0.1, "cm"))+
  theme(legend.text = element_text(size=8))
```
</br>

#### Merge plots (overall/fuel type and sector)
```{r tidy=TRUE}
sector_fuel_party <-ggarrange(co2_fuel, sector_p, ncol = 1, labels = c("", "c"), hjust = -3, vjust= 1.5, nrow = 2,font.label = list(size = 8, color = "black", face = "bold"))

co2_fuelsector<-annotate_figure(sector_fuel_party, left = text_grob("Weighted per capita CO2 (metric tons)", size=10, vjust = 2, rot = 90))
```
</br>

#### Output plot by total/fuel type and sector
```{r fig.height = 5, fig.width = 7, fig.align = "center"}
file.plot <- file.path(dir.plot, "co2_sector_fuel.png")
png(file.plot, height=5, width=7, units="in", bg="white", res=300)
co2_fuelsector
dev.off()
```

#### US State Map and state abbreviation
```{r tidy=TRUE}
names(state.abb) <- state.name
state.abbr <-data.frame(state.abb)
state_a <-setNames(cbind(rownames(state.abbr), state.abbr, row.names = NULL), c("State", "State_Abb"))

## Data for 1997-2021
dat_map <- dat_p %>% rename(state=State)
state_97 <- dat_map %>% filter(Year==1997)
state_19 <- dat_map %>% filter(Year==2019)
state_20 <- dat_map %>% filter(Year==2020)
state_21 <- dat_map %>% filter(Year==2021)
```

#### Map of per capita CO2: 1997
```{r tidy=TRUE}
state_97 <- state_97 %>% mutate(pco= log(pc_CO2))

co_97 <- plot_usmap(data = state_97, values = "pco", color = "grey")+ 
scale_fill_gradientn(colours = c("blue","white", "red"), 
  limits=c(1.96, 4.795), breaks = c(2.0, 2.5, 3.0, 3.5, 4.0, 4.4, 4.75), label = scales::comma) + 
  ggtitle("1997")+
  theme(plot.title = element_text(size = 8))+
  theme(plot.title = element_text(hjust = 0.5)) +    # Center ggplot title
  theme(legend.title = element_blank())+
  theme(legend.text = element_text(size=7))+
  theme(legend.position = "none", legend.margin=margin(t=-50)) +
  theme(legend.box = "vertical")+
  theme(legend.key.height = unit(0.3, 'cm'))+
  theme(legend.key.width = unit(1, 'cm'))
```

</br>
#### Map of per capita CO2: 2019
```{r tidy=TRUE}
state_19 <- state_19 %>% mutate(pco= log(pc_CO2))
co_19 <- plot_usmap(data = state_19, values = "pco", color = "grey")+ 
scale_fill_gradientn(colours = c("blue","white", "red"), 
  limits=c(1.96, 4.795), breaks = c(2.0, 2.5, 3.0, 3.5, 4.0, 4.4, 4.75),label = scales::comma) + 
  ggtitle("2019")+
  theme(plot.title = element_text(size = 8))+
  theme(plot.title = element_text(hjust = 0.5)) +
  theme(legend.title = element_blank())+
  theme(legend.text = element_text(size=7))+
  theme(legend.position = "bottom", legend.margin=margin(t=-50)) +
  theme(legend.box = "vertical")+
  theme(legend.key.height = unit(0.3, 'cm'))+
  theme(legend.key.width = unit(1, 'cm'))
```

</br>
#### Map of Per capita CO2: 2020
```{r tidy=TRUE}
state_20 <- state_20 %>% mutate(pco= log(pc_CO2))
co_20 <- plot_usmap(data = state_20, values = "pco", color = "grey")+ 
  scale_fill_gradientn(colours = c("blue","white","red"), 
  limits=c(1.96, 4.795), breaks = c(2.0, 2.5, 3.0, 3.5, 4.0, 4.4, 4.75), label = scales::comma) +
  ggtitle("2020")+
  theme(plot.title = element_text(size = 8))+
  theme(plot.title = element_text(hjust = 0.5)) +
  theme(legend.title = element_blank())+
  theme(legend.text = element_text(size=7))+
  theme(legend.direction = "horizontal") +
  theme(legend.position = "bottom", legend.margin=margin(t=-50)) +
  theme(legend.box = "vertical")+
  theme(legend.key.height = unit(0.3, 'cm'))+
  theme(legend.key.width = unit(1, 'cm'))
```

</br>
#### Map of Per capita CO2: 2021
```{r tidy=TRUE}
state_21 <- state_21 %>% mutate(pco= log(pc_CO2))
co_21 <- plot_usmap(data = state_21, values = "pco", color = "grey")+ 
  scale_fill_gradientn(colours = c("blue","white","red"), 
    limits=c(1.96, 4.795), breaks = c(2.0, 2.5, 3.0, 3.5, 4.0, 4.4, 4.75), label = scales::comma) +
  ggtitle("2021")+
  theme(plot.title = element_text(size = 8))+
  theme(plot.title = element_text(hjust = 0.5)) +
  theme(legend.title = element_blank())+
  theme(legend.text = element_text(size=7))+
  theme(legend.direction = "horizontal") +
  theme(legend.position = "bottom", legend.margin=margin(t=-50)) +
  theme(legend.box = "vertical")+
  theme(legend.key.height = unit(0.3, 'cm'))+
  theme(legend.key.width = unit(1, 'cm'))
```

</br>
#### Combine all US Maps of CO2 emissions 1997, 2019, 2020 and 2021
```{r fig.height = 3, fig.width = 7, fig.align = "center"}

legend <- get_legend(co_19)
          
co_p <- ggarrange(co_97, co_19, co_20, co_21, labels = "a", vjust=7, hjust = -2, font.label = list(size = 8, 
        color = "black", face = "bold"), ncol = 4, nrow = 1, legend="none", common.legend=F)+
        theme(plot.margin = unit(c(0, 5.5, 0, 5.5), "pt"))

co_le <- grid.arrange(co_p, legend, nrow = 2, layout_matrix = rbind(1, c(NA, 2, NA)), heights = c(3, -0.3))

co_pl <- annotate_figure(co_le, bottom = text_grob("Log of per capita CO2 (metric tons)", size=8))+
  theme(plot.margin = unit(c(0, 5.5, 0, 5.5), units= "pt"))
```


#### Map of state PVG: 1997, 2019, 2020 and 2021
```{r tidy=TRUE}
pv_97 <- plot_usmap(data = state_97, values = "PVG", color = "grey")+ 
  scale_fill_gradientn(colours = c("red","white", "blue"), limits=c(-51.42, 38.01), breaks = seq(-50, 38, 12), 
           name = "Partisan voting gap", label = scales::comma) + 
  theme(legend.title = element_blank())+
  theme(legend.text = element_text(size=7))+
  theme(legend.direction = "horizontal") +
  theme(legend.position = "bottom", legend.margin=margin(t=-30)) +
  theme(legend.box = "vertical")+
  theme(legend.key.height = unit(0.3, 'cm'))+
  theme(legend.key.width = unit(1, 'cm'))

pv_19 <- plot_usmap(data = state_19, values = "PVG", color = "grey")+ 
  scale_fill_gradientn(colours = c("red","white","blue"), limits=c(-51.42, 38.01), breaks = seq(-50, 38, 12), 
        name = "Partisan voting gap", label = scales::comma) + 
  theme(legend.title = element_blank())+
  theme(legend.text = element_text(size=7))+
  theme(legend.direction = "horizontal") +
  theme(legend.position = "bottom", legend.margin=margin(t=-30)) +
  theme(legend.box = "vertical")+
  theme(legend.key.height = unit(0.3, 'cm'))+
  theme(legend.key.width = unit(1, 'cm'))

pv_20 <- plot_usmap(data = state_20, values = "PVG", color = "grey")+ 
  scale_fill_gradientn(colours = c("red","white","blue"), limits=c(-51.42, 38.01), breaks = seq(-50, 38, 12), 
         name = "Partisan voting gap", label = scales::comma) + 
  theme(legend.title = element_blank())+
  theme(legend.text = element_text(size=7))+
  theme(legend.direction = "horizontal") +
  theme(legend.position = "none", legend.margin=margin(t=-30)) +
  theme(legend.box = "vertical")+
  theme(legend.key.height = unit(0.3, 'cm'))+
  theme(legend.key.width = unit(1, 'cm'))

pv_21 <- plot_usmap(data = state_21, values = "PVG", color = "grey")+ 
  scale_fill_gradientn(colours = c("red","white","blue"), limits=c(-51.42, 38.01), breaks = seq(-50, 38, 12), 
          name = "Partisan voting gap", label = scales::comma) + 
  theme(legend.title = element_blank())+
  theme(legend.text = element_text(size=7))+
  theme(legend.direction = "horizontal") +
  theme(legend.position = "none", legend.margin=margin(t=-30)) +
  theme(legend.box = "vertical")+
  theme(legend.key.height = unit(0.3, 'cm'))+
  theme(legend.key.width = unit(1, 'cm'))

```

</br>

#### Combine Maps of state PVG: 1997, 2019, 2020 and 2021
```{r fig.height = 3, fig.width = 7, fig.align = "center"}
legend <- get_legend(pv_19)
          
pv_p <- ggarrange(pv_97, pv_19, pv_20, pv_21, labels = "b", vjust=7, hjust=-2,font.label = list(size = 8, color = "black", face = "bold"), ncol = 4, nrow = 1, common.legend=F, legend="none")
        theme(plot.margin = unit(c(0, 5.5, 0, 5.5), "pt"))

pv_le <- grid.arrange(pv_p, legend, nrow = 2, layout_matrix = rbind(1, c(NA, 2, NA)), heights = c(3, -0.3))
  
pv_pl <- annotate_figure(pv_le, bottom = text_grob("Partisan voting gap (PVG)", size=8))+
        theme(plot.margin = unit(c(0, 5.5, 0, 5.5), "pt"))
```
</br>

#### Naive across-state correlation between PVG and CO2 emissions
```{r tidy=TRUE}
dat_1997 <- dat_p %>% filter(Year==1997)
dat_2019 <- dat_p %>% filter(Year==2019)
dat_2020 <- dat_p %>% filter(Year==2020)
dat_2021 <- dat_p %>% filter(Year==2021)

wtd.cor(log(dat_1997$pc_CO2), dat_1997$PVG, weight=dat_1997$pop_mi)
wtd.cor(log(dat_2019$pc_CO2), dat_2019$PVG, weight=dat_2019$pop_mi)
wtd.cor(log(dat_2020$pc_CO2), dat_2020$PVG, weight=dat_2020$pop_mi)
wtd.cor(log(dat_2021$pc_CO2), dat_2021$PVG, weight=dat_2021$pop_mi)
```

#### Scattor plots: PVG versus CO2
```{r fig.height = 3, fig.width = 7, fig.align = "center"}
s97 <- ggplot(dat_1997, aes(x=PVG, y = log(pc_CO2)))+
  geom_point(aes(color=Party, size = pop_mi), shape=21)+
  geom_text(aes(label = State.Abb), hjust = 0.5,  vjust = -1, size = 2) +
  scale_color_manual(labels = c("Blue states", "Red states"),
                     values=c("blue", "red"), guide = guide_legend(reverse=T))+
  geom_smooth(mapping=aes(weight=pop_mi), method=lm, se=T, color="black")+
  geom_smooth(method=lm, se=T, linetype="dotted", color="black")+
  labs(x="", y = "")+
  scale_y_continuous(limits = c(1, 5), breaks = seq(1, 5, 1)) +
  theme(legend.title = element_blank())+
  geom_text(x=-8, y=1.5, label="r = - 0.573 \n p = 1.36e-05", size=2.5)+
  geom_label(x= 0, y=4.9, label="1997", size =2.5)+
  theme(legend.position = c(0.9, 0.8), legend.text = element_text(size=8))+
  theme(axis.text=element_text(size=6))

s19 <- ggplot(dat_2019, aes(x=PVG, y = log(pc_CO2)))+
  geom_point(aes(color=Party, size=pop_mi), shape =21)+
  geom_text(aes(label = State.Abb), hjust = 0.5,  vjust = -1, size = 2) +
  scale_color_manual(labels = c("Blue states", "Red states"),
                     values=c("blue", "red"), guide = guide_legend(reverse=T))+
  geom_smooth(mapping=aes(weight=pop_mi), method=lm, se=T, color="black")+
  geom_smooth(method=lm, se=T, linetype="dotted", color="black")+
  labs(x="", y = "")+
  scale_y_continuous(limits = c(1, 5), breaks = seq(1, 5, 1)) +
  theme(axis.title.x = element_text(size = 8))+
  theme(legend.title = element_blank())+
  geom_text(x=-10, y=1.5, label="r = - 0.770 \n p = 6.39e-11", size=2.5)+
  geom_label(x= -5, y=4.9, label="2019", size =2.5)+
  theme(legend.position = c(0.9, 0.8), legend.text = element_text(size=8))+
  theme(axis.text=element_text(size=6))

s20 <- ggplot(dat_2020, aes(x=PVG, y = log(pc_CO2)))+
  geom_point(aes(color=Party, size=pop_mi), shape=21)+
  geom_text(aes(label = State.Abb), hjust = 0.5,  vjust = -1, size = 2) +
  scale_color_manual(labels = c("Blue states", "Red states"),
                     values=c("blue", "red"), guide = guide_legend(reverse=T))+
  geom_smooth(mapping=aes(weight=pop_mi), method=lm, se=T, color="black")+
  geom_smooth(method=lm, se=T, linetype="dotted", color="black")+
  labs(x="", y = "")+
  scale_y_continuous(limits = c(1, 5), breaks = seq(1, 5, 1)) +
  geom_text(x=-10, y=1.5, label="r = - 0.796 \n p = 5.07e-12", size=2.5)+
  geom_label(x=-5, y=4.9, label="2020", size =2.5)+
  theme(legend.title = element_blank())+
  theme(legend.position = c(0.9, 0.8),legend.text = element_text(size=8))+
  theme(axis.text=element_text(size=6))

s21 <- ggplot(dat_2021, aes(x=PVG, y = log(pc_CO2)))+
  geom_point(aes(color=Party, size=pop_mi), shape=21)+
  geom_text(aes(label = State.Abb), hjust = 0.5,  vjust = -1, size = 2) +
  scale_color_manual(labels = c("Blue states", "Red states"),
                     values=c("blue", "red"), guide = guide_legend(reverse=T))+
  geom_smooth(mapping=aes(weight=pop_mi), method=lm, se=T, color="black")+
  geom_smooth(method=lm, se=T, linetype="dotted", color="black")+
  labs(x="", y = "")+
  scale_y_continuous(limits = c(1, 5), breaks = seq(1, 5, 1)) +
  geom_text(x=-10, y=1.5, label="r = - 0.783 \n p = 1.79e-11", size=2.5)+
  geom_label(x=-5, y=4.9, label="2021", size =2.5)+
  theme(legend.title = element_blank())+
  theme(legend.position = c(0.9, 0.8),legend.text = element_text(size=8))+
  theme(axis.text=element_text(size=6))

splot <-ggarrange(s97, s19, s20, s21, labels = c("c", "", "", ""), vjust = 0, hjust = -2, 
                        font.label = list(size = 8, color = "black", face = "bold"), 
                        ncol = 4, nrow = 1, legend = "none")
splot <- annotate_figure(splot, bottom = text_grob("Partition voting gap (PVG)", size=8, vjust = -2))+
      theme(plot.margin = unit(c(5.5, 0, 5.5, 0), "pt"))
```

#### All plots: maps and correlation
```{r fig.height = 7, fig.width = 7, fig.align = "center"}
co_pv <- cowplot::plot_grid(co_pl, pv_pl, splot, nrow = 3, ncol = 1)

file.plot <- file.path(dir.plot, "co_pv.png")
png(file.plot, height=7, width=7, units="in", bg= "white", res=300)
co_pv
dev.off()
```

#### Weighted mean per capita CO2 by HDD, CDD, income, price, and expenditure
```{r tidy=TRUE}

mean_party <- dat_p %>% group_by(Year, Party) %>% 
            summarise(party_CO2 = weighted.mean(pc_CO2, pop_mi),
            party_price = weighted.mean(price, pop_mi), 
            party_expnd = weighted.mean(expend, pop_mi),
            party_gdp = weighted.mean(pc_rgdp, pop_mi), 
            party_PVG = weighted.mean(PVG, pop_mi))
mean_HDD <- dat_p %>% group_by(Year, HDD) %>% 
          summarise(HDD_CO2 = weighted.mean(pc_CO2, pop_mi),
          HDD_price = weighted.mean(price, pop_mi), 
          HDD_expnd = weighted.mean(expend, pop_mi),
          HDD_gdp = weighted.mean(pc_rgdp, pop_mi), 
          HDD_PVG = weighted.mean(PVG, pop_mi), n())
                                                        
mean_CDD <- dat_p %>% group_by(Year, CDD) %>% 
        summarise(CDD_CO2 = weighted.mean(pc_CO2, pop_mi),
        CDD_price = weighted.mean(price, pop_mi), 
        CDD_expnd = weighted.mean(expend, pop_mi),
        CDD_gdp = weighted.mean(pc_rgdp, pop_mi), 
        CDD_PVG = weighted.mean(PVG, pop_mi), n())
mean_HL <- do.call(bind_cols, lapply(list(mean_party, mean_HDD, mean_CDD), `[`, ))
## output data
file.output <- file.path(dir.outdata, "mean_HL.csv")
write.table(mean_HL, file.output, sep=",", row.names=FALSE, col.names=TRUE, quote=FALSE)
```

#### Number of countries by HDD, CDD, income, price, geography and expenditure
```{r tidy=TRUE}
nparty <- dat_p %>% group_by(Year, Party) %>% summarise(n())
nHDD <- dat_p %>% group_by(Year, HDD) %>% summarise(n())
nCDD <- dat_p %>% group_by(Year, CDD) %>% summarise(n())
nLCV <- dat_p %>% group_by(Year, HL_LCV) %>% summarise(n())
nexp <- dat_p %>% group_by(Year, HL_Expend) %>% summarise(n())
ngdp <- dat_p %>% group_by(Year, in_class) %>% summarise(n())
ngeo <- dat_p %>% group_by(Year, Geography) %>% summarise(n())

num_HL <- do.call(bind_cols, lapply(list(nparty, nHDD, nCDD, nLCV, nexp, ngdp, ngeo), `[`, ))
## output data
file.output <- file.path(dir.outdata, "num_HL.csv")
write.table(num_HL, file.output, sep=",", row.names=FALSE, col.names=TRUE, quote=FALSE)
```
